Ensure access to PDFDocument is always concurrency safe (#58)
# Ensure access to PDFDocument is always concurrency safe

## :recycle: Current situation & Problem
The PR #51 added several interactions with `PDFDocument` that cannot be
proven to be concurrency safe. This makes the package incompatible with
Swift 6 and unable to compile. The issue was tracked in #57.
This PR attempts to resolve these issues. However, this not possible
without breaking existing API. So see the initial state of the PR as a
starting point for discussion on how we version this change and how we
provide compatibility with Swift 5.10 environment as most of the changes
require Swift 6 features.

`PDFDocument` is a
non-`[Sendable`](https://developer.apple.com/documentation/swift/sendable)
type (and probably won't be ever as it is an open class type inheriting
from NSObject, see [Sendable
Classes](https://developer.apple.com/documentation/swift/sendable#Sendable-Classes)).
However, non-Sendable types can cross actor boundaries when passed as
[`sending`](https://github.com/swiftlang/swift-evolution/blob/main/proposals/0430-transferring-parameters-and-results.md)
parameters. This keyword ensures that the caller won't maintain any
references and ownership is fully transferred to the callee (for
non-Sendable types). As the `PDFDocument` is just a result type we mean
to transfer to the `Standard` that implements the `ConsentConstraint`
and don't require to maintain ownership within SpeziOnboarding, we can
easily send the PDFDocument to the Standard and provide ownership to the
Standard.

EDIT: Here is a great deep dive into non-copyable types: [Consume
noncopyable types in Swift](https://developer.apple.com/wwdc24/10170).

closes #47
closes #57

## :gear: Release Notes 
All changes are annotated if they require Swift 6 language features and
if they are considered a breaking change. Note that we assume framework
users to operate in Swift 5 mode. As Swift 6 isn't released yet, we do
not consider any changes to concurrency annotations that do not create
compiler errors in Swift 5 (even with strict concurrency checking) as a
breaking change.

The following changes were made according to the approach explained
above:
* The legacy `OnboardingConstraint` receives its `PDFDocument` as a
`sending` parameter. This **requires Swift 6** and is a **breaking
change**.
* The `OnboardingDataSource` receives the `PDFDocument` as a `sending`
parameter. This **requires Swift 6** and is **not** a breaking change.
* `ConsentDocumentExport` was turned into a non-Copyable type that is
non-Sendable. The access to the PDF is `consuming` and returns the
`PDFDocument` as a `sending` result. This **requires Swift 6** and is a
**breaking change**.
* `consuming` can only be applied to `func` declarations and `sending`
only to results types of `func` declarations. Therefore, we needed to
replace the `async` `pdf` property.
* We currently keep the `consumePDF` method async as the documentation
suggest that this might be required in the future. Is that really
something that will come?
* The `ConsentConstraint` was updated to receive the
`ConsentDocumentExport` as a `sending` and `consuming` parameter. This
**requires Swift 6** and is a **breaking change**.


## :books: Documentation
_TBA_


## :white_check_mark: Testing
The test app was updated and existing unit tests will cover the
refactored code.


## :pencil: Code of Conduct & Contributing Guidelines 

By submitting creating this pull request, you agree to follow our [Code
of
Conduct](https://github.com/StanfordSpezi/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordSpezi/.github/blob/main/CONTRIBUTING.md):
- [x] I agree to follow the [Code of
Conduct](https://github.com/StanfordSpezi/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordSpezi/.github/blob/main/CONTRIBUTING.md).

---------

Co-authored-by: Paul Schmiedmayer <PSchmiedmayer@users.noreply.github.com>
Co-authored-by: Philipp Zagar <zagar@stanford.edu>