Improve OnboardingStack; some other changes (#67)
# Improve OnboardingStack; some other changes

## :recycle: Current situation & Problem
- OnboardingStack
- Added the ability to optionally pass an externally-managed
`OnboardingNavigationPath`
- The reason this exists is to better deal with situations where the
entire view hierarchy, or large parts of it, are reloaded while an
OnboardingStack is presented (eg in response to an `Observable` state
changing somewhere furter up in the hierarchy. If the OnboardingStack
uses a path owned by the stack itself, the path will get reset as part
of the view reload, and the current position in the stack will get lost.
Injecting an externally-owned path allows us to work around this and
maintain the state of the OnboardingStack.
- (The specific use case here is conditionally loading Spezi modules as
part of an onboarding step. Since the `Spezi` class itself is
observable, loading a new module will cause all views dependent on it
(ie, typically the entire app) to get reloaded. We need a way of being
able to do deferred module loading without it also resetting the entire
onboarding progress.)
- The OnboardingStack now sets a `isInOnboardingStack` environment value
to true on all views contained within the stack.
- OnboardingNavigationPath:
- now uses an `[OnboardingStepIdentifier]` instead of a SwiftUI
`NavigationPath`
- OnboardingStepIdentifier:
- now uses `String(reflecting:)` instead of `String(describing:)` when
deriving identifiers from view types.
- the difference is that describing won't use the full typename, whereas
reflecting will.
- this fixes a bug where two different view types that have the same
name but are nested in different types would have been treated as being
identical by SpeziOnboarding.
- OnboardingView:
- added `disablePadding(_:)` modifier function, which allows specifying
which edges of the view should not get implicit padding
- apart from that, the OnboardingView will now also disable its top
padding when it is contained in an OnboardingStack. The reason for this
is that the OnboardingStack will already put the views into a
NavigationStack (which in this context will typically have an empty
navigation title), which will act as a source of padding. Adding the
OnboardingView's extra top padding is redundant and pushes the content
(and eg the OnboardingView's title) far too much down.
- Consent:
- Removed a bunch of duplicate code and rewrote it to be a bit more
elegant. Instead of having separate definitions of whole structs for
UIKit and AppKit (ie, based on whether `UIFont` or `NSFont`, or
`UIColor` or `NSColor` is available), we now have simple typealiases for
the two and can use a single definition for anything else.
- Other:
- Moved the test PDFs out of LFS, since that caused some issues with SPM
    - All of these changes should be backwards-compatible

## :books: Documentation
New code is documented

## :white_check_mark: Testing
WIP

## :pencil: Code of Conduct & Contributing Guidelines
By creating and submitting this pull request, you agree to follow our
[Code of
Conduct](https://github.com/StanfordSpezi/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordSpezi/.github/blob/main/CONTRIBUTING.md):
- [x] I agree to follow the [Code of
Conduct](https://github.com/StanfordSpezi/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordSpezi/.github/blob/main/CONTRIBUTING.md).

---------

Signed-off-by: Lukas Kollmer <hey@lukaskollmer.de>